adde pre and post phases for planning optimizer
